Rename nonlinearities to neurons
While they aren't all necessarily neuron types,
they will essentially be treated like neurons in that
they can act as Ensemble.neurons.
