+++ nengo/neurons.py
