+++ nengo/tests/test_neurons.py
